deploy/requirements.txt
python-dotenv
requests
openai
img2pdf
arxiv
feedparser
